API: Move PluginMainContainer to its own module
